Remove dependency on Laravel Application
