+++ src/Illuminate/Console/Application.php
@@ -82,6 +82,8 @@ class Application extends SymfonyApplication implements ApplicationContract
     {
         if ($command instanceof Command) {
             $command->setLaravel($this->laravel);
+        } elseif ($command instanceof BaseCommand) {
+            $command->setContainer($this->laravel);
         }
 
         return $this->addToParent($command);
@@ -157,7 +159,7 @@ class Application extends SymfonyApplication implements ApplicationContract
     /**
      * Get the Laravel application instance.
      *
-     * @return \Illuminate\Contracts\Foundation\Application
+     * @return \Illuminate\Contracts\Container\Container
      */
     public function getLaravel()
     {
